# [WIP] Speed up MinHash by using a vector instead of a set