Keep the list of N minimum hash values dense and cache local by using a vector instead of a set.

Thoughts on the general sentiment?

I used `time_add_sequence` and `time_add_hash` from #100 as benchmarks:

| bench | master [s] | this PR [s] |
|----------|-----------|------------|
| add_sequence | 0.81 | 0.68 |
| add_hash | 0.16 | 0.11 |

There are a few tests that don't yet pass and we can be much smarter in the `combine` methods. Before spending time cleaning it up: is the performance boost worth it? is the benchmark I used useful or should I try something else to conclude that it does speed things up?

- [ ] Is it mergeable?
- [ ] `make test` Did it pass the tests?
- [ ] `make coverage` Is the new code covered?
- [ ] Did it change the command-line interface? Only additions are allowed
  without a major version increment. Changing file formats also requires a
  major version number increment.
- [ ] Was a spellchecker run on the source code and documentation after
  changes were made?