@ctb 2017-01-12
this sounds excellent, and even ~10-15% speedups are very welcome at this
point.

question: if you have a large (many thousands) collection of hashes,
what kind of speedup do you get on on insertion?  this is relevant to
a few use cases that I've been working with.