@betatim 2017-01-12
(For the historians: I ended up with this after [experimenting](https://gist.github.com/betatim/6fc927f6d2dfa09d7f5f84b3244626b8) with various ideas. It is interesting to look at how things [scale as size of the list you maintain](https://gist.github.com/betatim/6fc927f6d2dfa09d7f5f84b3244626b8#file-results-txt))