@betatim 2017-01-12
You want to insert 1000 hashes in one go? Or you want to have a MinHash with 1000s of "smallest hash values"?