@ctb 2017-01-12
the latter -- well, both, actually, but I'm wondering how well the
"1000s of smallest hash values" would perform.