@ctb 2017-01-12
wow, so, bench_vector_insert seems like the way to go, eh?