@betatim 2017-01-12
Another idea worth trying: when you call `add_sequence` you can not observe the minimum hashes until it has completed. Maybe we can save time by `push_back`'ing "a lot of hashes" and only after a large number of them sort and shrink the vector used to store them. Right now we sort and shrink after each hash has been inserted.