@lgautier 2017-01-12
I have not looked at the details of the benchmark (I assume it is building a MinHash sketch from a sequence), nor whether the benchmark results are the results of repeated measurement) but the speed gain seems worthy.

Speaking of speed-up, I have started thinking about design for very general library for MinHash with flexibility and ease-of-experimentation in mind. Clean slate, just starting from what MASH and sourmash_lib are doing, the needs of the prospective JSON exchange format, and what kinds of patterns would make implementing the ideas I want to try with MinHashes easy to implement.

I get surprising results when I tried benchmarking how bad was the performance loss (only-python, C-extension code only doing calculation of hash values): initiai tests suggest that this new design & implementation is 3 to 4 times faster than all-C `sourmash_lib._minhash`, and I have hints that this new implementation might scale better (I am currently looking at that more in details). For the time being I threw it into a jupyter notebook: https://github.com/lgautier/mashing-pumpkins/blob/master/doc/notebooks/MinHash%2C%20design%20and%20performance.ipynb