@betatim 2017-01-12
> wow, so, bench_vector_insert seems like the way to go, eh?

yes, it is what I tried to put into this PR. Probably the gains are less than in the benchmark because we can't nicely line up many hashes to be inserted in bulk, but do one at a time.