@betatim 2017-01-12
@lgautier should we start a new issue to discuss performance tricks, ideas and such? Feels like we will get a big mess if we discuss that here, as well as review the code :-/