@lgautier 2017-01-12
@betatim : Most definitely. No intent to discuss further how and why here. Sorry if not clear: the message was that if performances are the aim, there appear to remain room for other ways than C/C++ + close-to-metal tricks (and I assumed not main would believe that without a concrete demonstration ;-)).