@betatim 2017-01-13
I'll stop tinkering now. :grinning: I'll add some numbers on how things scale as a function of the number of minimum hashes you keep track of.

Ready for review! @luizirber or @ctb