@luizirber 2017-01-14
@betatim using the vector also opens the path to eventually use a priority_queue again, I initially implemented the track abundance feature using it but figured that just keeping track of the max hash was enough at that point: https://github.com/dib-lab/sourmash/pull/49/commits/53e5e2f2bedff1eccec328e92dfef01adfe7b046
(and I think a big part of the performance @lgautier is getting from mashing pumpkins is being smart of heaps, right? =])