@lgautier 2017-01-14
> (and I think a big part of the performance @lgautier is getting from mashing pumpkins is being smart of heaps, right? =])

No shroud of mystery about how this is implemented as the code is just there for all to see: I do use a heap. No other claim ;-).

By now I probably made my point that algorithms > (language | close-to-hardware), at least for a first look at thing... but that's likely not the only element responsible for the difference in performance.

I have not unraveled everything but I have more for you, if interested. Here is a figure from a quick exploration I did. Each panel has:
- the length of the sequence for which a MinHash is build on the x-axis
- the speed (in MB/sec)
- the header of the panel tells the max_size of the MinHash (number of elements in the MinHash)
- for mashing-pumpkins I am exploring different batch sizes (calling the hashing function is in C, done by batch), each as a line graph

![image](https://cloud.githubusercontent.com/assets/650866/21951767/638209f6-d9da-11e6-9e0f-8490f8238296.png)

`sourmash`'s performance is pretty much uniform across the range of parameters tested. I do benefit of having larger input sequence to compute a MinHash for (may be there is a preemptive memory allocation scheme in Python that is responsible for that effect). On the other end, the performance is degrading as the size of the MinHash is increasing. With `max_size` 10,000 or 20,000 `sourmash` is faster... until the input sequence gets large enough to see the benefit mentioned right before counterbalance it.