@betatim 2017-01-15
AFAIK `priority_queue` and `make_heap` in the c++ stdlib don't deal with duplicate entries. So if you have (say) 5 in your list of minhashes and add it again, it will do that instead of doing nothing. Not sure what the best way of solving that is. The obvious way to solve it is to maintain a `set` of already seen values (or implement a `search_heap`). Unfortunately both `set` and `unordered_set` in the c++ stdlib have to be implemented in a way that means you end up chasing pointers.

For the record, `std::priority_queue` without checking for duplicates is faster (and less code) than d148a40.

---

@lgautier can you add the sanity checking that `sourmash` does to your python only implementation? The gap between `add_sequence` and `add_hash` in the above table is essentially all due to checking for weird characters, building the reverse complement, etc

It also tells us that ~20% of inserting a new hash into the minhash is about computing the hash, checking whether it needs inserting, and inserting it. Most of the time is spent elsewhere :grin: