@lgautier 2017-01-15
> @lgautier can you add the sanity checking that sourmash does to your python only implementation? The gap between add_sequence and add_hash in the above table is essentially all due to checking for weird characters, building the reverse complement, etc

Fair point. I did not see that. Stay tuned.