@lgautier 2017-01-16
@betatim 
> The obvious way to solve it is to maintain a `set` of already seen values (or implement a `search_heap`). Unfortunately both set and unordered_set in the c++ stdlib have to be implemented in a way that means you end up chasing pointers.

FWIW I am using a `set` (lookup in O(1) - a `search_heap` would not do better and would be yet-an-other-thing to implement).

I am working out the last design bits (around parallelization) and when done will see whether I can show in a PR with simple adapter code that it could fit in.