@betatim 2017-01-17
> FWIW I am using a `set` (lookup in O(1) - a `search_heap` would not do better and would be yet-an-other-thing to implement).

Jupp, in python both `dict` and `set` are O(1), but in c++ unfortunately they aren't (so smart).