@betatim 2017-01-17
My summary: I would merge this now, then see if we can improve on it some more later. For example we could use a non stdlib set (keyword: open addressing) to get O(1) lookup and insertion.

👍 on data structures beat languages, however I am also convinced if you take the smart algorithm setup from python and move to c++ it has to be as fast or faster. If it is slower there must be a bug somewhere ;) (technical debt/complexity being a whole different conversation)